[objc-appscript] added workaround for OS X 10.6's return ID bug; cleaned up static analysis warnings
git-svn-id: https://appscript.svn.sourceforge.net/svnroot/appscript@705 5e899d65-cff2-4c64-b955-3fa930e4d869

+++ objc-appscript/trunk/src/Appscript/base.m
@@ -27,7 +27,7 @@
 }
 
 
-- (unsigned long)hash {
+- (NSUInteger)hash {
 	if (!cachedHash)
 		cachedHash = [[self description] hash];
 	return cachedHash;
